Finished first example with using business layer

+++ Group2DataDriven/DataLayer/AuthorRepository.cs
@@ -12,15 +12,16 @@ namespace DataLayer
     {
         private SqlConnection connection;
 
-        public AuthorRepository()
+        public AuthorRepository(string connectionString)
         {
-            connection = new SqlConnection(@"Server=(localDB)\.\Kentico;Database=Scify2;Trusted_Connection=True;");
+            //TODO: connection string should not be hardcoded
+            connection = new SqlConnection(connectionString);
             connection.Open();
         }
 
         public Author GetAuthorByName(string name)
         {
-            using (SqlCommand cmd = new SqlCommand("select top 1 * from authors where Name like '%' + @name + '%'", connection))
+            using (SqlCommand cmd = new SqlCommand("select * from authors where Name=@name", connection))
             {
                 cmd.Parameters.AddWithValue("@name", name);
 
@@ -32,20 +33,7 @@ namespace DataLayer
                         return null;
                     }
 
-                    var resultId = (int)reader["ID"];
-                    var resultName = (string)reader["Name"];
-                    var birthDate = (DateTime)reader["DateOfBirth"];
-                    var deathDate = reader["DateOfDeath"] as DateTime?;
-                    
-                    var author = new Author
-                    {
-                        ID = resultId,
-                        Name = resultName,
-                        BirthDate = birthDate,
-                        DeathDate = deathDate
-                    };
-
-                    return author;
+                    return GetAuthorFromDataReader(reader);
                 }
                 
             }
@@ -54,17 +42,45 @@ namespace DataLayer
 
         public bool DeleteAuthor(int id)
         {
-            throw new NotImplementedException();
+            using (var cmd = new SqlCommand("delete from authors where ID=@id", connection))
+            {
+                cmd.Parameters.AddWithValue("@id", id);
+
+                return cmd.ExecuteNonQuery() > 0;
+            }
         }
 
         public bool DeleteAuthor(Author author)
         {
+            if (author != null && author.ID == 0)
+            {
+                author = GetAuthorByName(author.Name);
+            }
+
+            if (author == null)
+            {
+                return false;
+            }
+            
             return DeleteAuthor(author.ID);
         }
 
         public IEnumerable<Author> GetAllAuthors()
         {
-            throw new NotImplementedException();
+            var authors = new List<Author>();
+
+            using (var cmd = new SqlCommand("select * from authors", connection))
+            {
+                using (var reader = cmd.ExecuteReader())
+                {
+                    while (reader.Read())
+                    {
+                        authors.Add(GetAuthorFromDataReader(reader));
+                    }
+                }
+            }
+
+            return authors;
         }
 
         public Author GetAuthor(int id)
@@ -81,34 +97,73 @@ namespace DataLayer
                         return null;
                     }
 
-                    var resultId = (int)reader["ID"];
-                    var name = (string)reader["Name"];
-                    var birthDate = (DateTime)reader["DateOfBirth"];
-                    var deathDate = reader["DateOfDeath"] as DateTime?;
-                    
-                    var author = new Author
-                    {
-                        ID = resultId,
-                        Name = name,
-                        BirthDate = birthDate,
-                        DeathDate = deathDate
-                    };
-
-                    return author;
+                    return GetAuthorFromDataReader(reader);
                 }
 
             }
             
         }
 
+        private static Author GetAuthorFromDataReader(SqlDataReader reader)
+        {
+            var resultId = (int)reader["ID"];
+            var name = (string)reader["Name"];
+            var birthDate = (DateTime)reader["DateOfBirth"];
+            var deathDate = reader["DateOfDeath"] as DateTime?;
+
+            var author = new Author
+            {
+                ID = resultId,
+                Name = name,
+                BirthDate = birthDate,
+                DeathDate = deathDate
+            };
+
+            return author;
+        }
+
         public Author SaveAuthor(Author author)
         {
-            throw new NotImplementedException();
+            if (author.ID != 0)
+            {
+                return UpdateAuthor(author);
+            }
+
+            if (ExistByName(author))
+            {
+                throw new ArgumentException("Author with that name exists!");
+            }
+
+            using (var cmd = new SqlCommand("insert into Authors (Name, DateOfBirth, DateOfDeath) values(@name, @birthDate, @deathDate)", connection))
+            {
+                cmd.Parameters.AddWithValue("@name", author.Name);
+                cmd.Parameters.AddWithValue("@birthDate", author.BirthDate);
+                cmd.Parameters.AddWithValue("@deathDate", author.DeathDate);
+
+                cmd.ExecuteNonQuery();
+
+                return GetAuthorByName(author.Name);
+            }
         }
 
         public Author UpdateAuthor(Author author)
         {
+            if (author.ID == 0)
+            {
+                return SaveAuthor(author);
+            }
+
             throw new NotImplementedException();
+        }
+
+        private bool ExistByName(Author author)
+        {
+            using (var cmd = new SqlCommand("select count(*) from authors where Name=@name", connection))
+            {
+                cmd.Parameters.AddWithValue("@name", author.Name);
+
+                return (int)cmd.ExecuteScalar() > 0;
+            }
         }
 
         #region IDisposable Support
@@ -148,6 +203,6 @@ namespace DataLayer
         }
         #endregion
 
-
+        
     }
 }
